[WIP] Add an option to print the intermediate representation
It's very useful to figure out what IR is generated to find out why things are not behaving the way they should.

The way this PR works is by going through the IR and flatten all the extra concat. Then, it prints them as a valid JavaScript string and uses prettier to actually properly indent it (which is extremely cool!).

Note: the integration with the command line is super hacky. I just wanted to put this out there so we can discuss before starting a refactor.

```js
echo "<div>&lt;</div>" | ./bin/prettier.js --stdin --debug-print-docs --single-quote
[
  [
    [
      [
        group(
          group([
            group([ '<', 'div', group([ indent(2, []), softline() ]), '>' ]),
            '&lt;',
            [ '</', 'div', '>' ]
          ]),
          [
            group([
              group([ '<', 'div', group([ indent(2, []), softline() ]), '>' ]),
              '&lt;',
              [ '</', 'div', '>' ]
            ]),
            group([
              group([ '<', 'div', group([ indent(2, []), softline() ]), '>' ]),
              indent(2, groupBreak([ hardline(), '&lt;' ])),
              hardline(),
              [ '</', 'div', '>' ]
            ])
          ]
        ),
        ';'
      ]
    ],
    hardline()
  ]
];
```