[WIP] Add rationale document
I think it would be great to have such a document. This way we don't have to explain the same things over and over again on the issues and it helps users and contributors know what they are getting into.

This is a very rough draft with a lot of empty content but I think it shows the idea.