[WIP] Preserve empty lines in arrays
The logic should be good but unfortunately the logic that finds if we are at the end of the line is completely busted. There's a potential fix in #358 so we'll wait and see if it fixes it here.

Putting it up for review so that we can discuss it.

Fixes #311