[JSX] Put `}` on the same level of indent as `{`
If the children are going to have a level of indentation, we should put a soft empty line. This way the `}` is going to be at the same level of indentation as `{`. Sadly, I don't think there's a way to express this with the IR, but since we're only inline so few nodes, I just put the two that will have this behavior.