Better indentation for binary expressions
Right now we always indent any binary expression group (a sequence of binary expressions of the same predecence) which leads to weird looking code, especially in if statements.

This new logic tries to make it better. There are now three cases where we need to indent things:

1) If we are in a comparison operator that is unlikely to be chained (eg ===)
2) If the parent expects a level of indentation (eg, var a = ...)
3) If the parent is a binary expression with a different precedence level (eg, && inside of ||)

The biggest observable change is the fact that it no longer indents the second and on conditions inside of an if statement.

Before:

```js
if (
  this.state.open &&
    this.props.value &&
    this.props.value[consts.inspected] === false
) {
  this.inspect();
}
```

After:

```js
if (
  this.state.open &&
  this.props.value &&
  this.props.value[consts.inspected] === false
) {
  this.inspect();
}
```

You can see how this runs on the Nuclide codebase: https://gist.github.com/vjeux/510742148b0a011d0375f23f260bf859

Fix #567