[RFC] Do not format template literals
We have a lot of issues with template literals that are improperly indented with prettier and shouldn't be. Maybe the best solution for this is to actually just give up and reprint it exactly the way the user entered it. I'm not super happy about this but well, it's worth a shot.