Introduce JSON-RPC
In order to be avoid the startup cost of prettier which makes editor integrations feel very slow, we can keep the process alive and communicate via JSON-RPC. This is a bit more work to do on the integration side but enables <50ms response time instead of >300ms, which makes a huge difference in practice.

I'm using vscode-jsonrpc implementation which has been battle-tested inside of vscode and doesn't have any dependencies.