Modify `singleQuote` option to allow single quotes in jsx
We use jsx with single quotes in our application code:

  <Icon name='search' />

When I first tried out prettier, I thought that the `singleQuote` would
control jsx attributes as well. But I found out that it didn't.

It's likely that we are an outlier, so I didn't want to force
single-quotes all over. The solution I went with was to allow different
quote styles in regular js code and in jsx. The `singleQuote`
(--single-quote) option now takes a string that can have the following
values:

  "none" (default) - use double quotes everywhere
  "all" - use single quotes everywhere
  "js" - use single quotes in js code, double quotes in jsx
  "jsx" - use double quotes in jsx code, double quotes in js

Inspired by 5d13102, I kept support for the boolean option. It comes
with a deprecation warning, so people should find out that they have to
update soon.

Fixes #1080