Flatten nested ternaries
This keeps coming up that the way prettier prints nested ternaries is not good. Here's an attempt at solving the situation. A good compromise is to
- keep printing ternaries the way they are but do not indent nested ones
- have a single group for chained ternaries, this way the last ones also break instead of awkwardly being fitting in a single line

The implementation is pretty straightforward, just look at the parent and print it differently if it's a ternary and we are in the second group.

Fixes #737