Add `astTransform` option
As suggested in https://github.com/prettier/prettier/issues/1684#issuecomment-303547552,
this adds a way for users to transform their AST before formatting,
without having to do so as a separate step before `prettier`. They
simply provide a function that takes an AST and returns an AST.

For example:

```js
require("./").format("// this won't show up in the output", {
  astTransform: () => require("./src/parser").parse("0", { parser: "babylon" })
});
```

The option is also available in the CLI, by specifying a path (resolved
relative to the current working directory) to a module that exports the
transform function. For example, given `transform.js` containing:

```js
module.exports = () =>
  require("./src/parser").parse("0", { parser: "babylon" });
```

the following is equivalent to the API example above:

```bash
echo "// this won't show up in the output" | ./bin/prettier.js --stdin --ast-transform transform.js
```